+++ CHANGELOG.md
@@ -66,4 +66,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documented ``warm_start`` option for pretraining the generator
 - Documented ``opt_g_kwargs`` and ``opt_d_kwargs`` for custom optimizer
   arguments
+- Added options to log gradient norms, learning rates and weight histograms for
+  improved training diagnostics